# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MailFlow AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: templates
    description: Email template operations
  - name: campaigns
    description: Campaign operations
  - name: email-logs
    description: Email log operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all email templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"

  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a single template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
        "404":
          description: Not found
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a single campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /campaigns/{id}/recipients:
    get:
      operationId: listRecipients
      tags: [campaigns]
      summary: List recipients for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of recipients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipient"
    post:
      operationId: addRecipients
      tags: [campaigns]
      summary: Add recipients to a campaign (bulk upload from CSV)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipientBulkInput"
      responses:
        "201":
          description: Recipients added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddRecipients201"

  /campaigns/{id}/send:
    post:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Start sending a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Send started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendCampaign200"

  /campaigns/{id}/retry:
    post:
      operationId: retryCampaign
      tags: [campaigns]
      summary: Retry failed emails in a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retry started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetryCampaign200"

  /campaigns/{id}/test-email:
    post:
      operationId: sendTestEmail
      tags: [campaigns]
      summary: Send a test email for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestEmailInput"
      responses:
        "200":
          description: Test email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendTestEmail200"

  /campaigns/{id}/stats:
    get:
      operationId: getCampaignStats
      tags: [campaigns]
      summary: Get sending stats for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignStats"

  /email-logs:
    get:
      operationId: listEmailLogs
      tags: [email-logs]
      summary: List all email logs with optional filters
      parameters:
        - name: campaignId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of email logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailLog"
    delete:
      operationId: clearAllEmailLogs
      tags: [email-logs]
      summary: Delete all email logs
      responses:
        "204":
          description: All logs deleted

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent email activity (last 30 logs)
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailLog"

  /dashboard/chart:
    get:
      operationId: getDashboardChart
      tags: [dashboard]
      summary: Get daily email volume for bar chart (last 7 days)
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    EmailTemplate:
      type: object
      required: [id, name, subject, body, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
          nullable: true
        createdAt:
          type: string

    EmailTemplateInput:
      type: object
      required: [name, subject, body]
      properties:
        name:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
          nullable: true

    EmailTemplateUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
          nullable: true

    Campaign:
      type: object
      required: [id, name, status, sentCount, failedCount, pendingCount, totalCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        templateId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, running, done, paused]
        sentCount:
          type: integer
        failedCount:
          type: integer
        pendingCount:
          type: integer
        totalCount:
          type: integer
        createdAt:
          type: string

    CampaignInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        templateId:
          type: ["integer", "null"]

    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        templateId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, running, done, paused]

    Recipient:
      type: object
      required: [id, campaignId, name, email, status]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        metadata:
          type: ["object", "null"]
          additionalProperties:
            type: string
        status:
          type: string
          enum: [pending, sent, failed]
        error:
          type: ["string", "null"]

    RecipientRow:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        metadata:
          type: ["object", "null"]
          additionalProperties:
            type: string

    RecipientBulkInput:
      type: object
      required: [recipients]
      properties:
        recipients:
          type: array
          items:
            $ref: "#/components/schemas/RecipientRow"

    AddRecipients201:
      type: object
      required: [count]
      properties:
        count:
          type: integer

    SendCampaign200:
      type: object
      required: [message]
      properties:
        message:
          type: string

    RetryCampaign200:
      type: object
      required: [message]
      properties:
        message:
          type: string

    SendTestEmail200:
      type: object
      required: [message]
      properties:
        message:
          type: string

    EmailLog:
      type: object
      required: [id, campaignId, recipientEmail, subject, status, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        campaignName:
          type: ["string", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientEmail:
          type: string
        recipientName:
          type: ["string", "null"]
        subject:
          type: string
        status:
          type: string
          enum: [pending, sent, failed]
        error:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string

    TestEmailInput:
      type: object
      required: [to]
      properties:
        to:
          type: string

    CampaignStats:
      type: object
      required: [total, sent, failed, pending]
      properties:
        total:
          type: integer
        sent:
          type: integer
        failed:
          type: integer
        pending:
          type: integer

    DashboardStats:
      type: object
      required: [totalSent, totalFailed, totalPending, totalCampaigns, totalTemplates]
      properties:
        totalSent:
          type: integer
        totalFailed:
          type: integer
        totalPending:
          type: integer
        totalCampaigns:
          type: integer
        totalTemplates:
          type: integer

    ChartDataPoint:
      type: object
      required: [date, sent, failed]
      properties:
        date:
          type: string
        sent:
          type: integer
        failed:
          type: integer