# lib/api-zod/src/generated/types/chartDataPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MailFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChartDataPoint {
  date: string;
  sent: number;
  failed: number;
}