# lib/api-zod/src/generated/types/emailLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MailFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailLogStatus } from './emailLogStatus';

export interface EmailLog {
  id: number;
  campaignId: number;
  /** @nullable */
  campaignName?: string | null;
  /** @nullable */
  recipientId?: number | null;
  recipientEmail: string;
  /** @nullable */
  recipientName?: string | null;
  subject: string;
  status: EmailLogStatus;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}