# lib/api-zod/src/generated/types/recipient.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MailFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecipientMetadata } from './recipientMetadata';
import type { RecipientStatus } from './recipientStatus';

export interface Recipient {
  id: number;
  campaignId: number;
  name: string;
  email: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  domain?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  metadata?: RecipientMetadata;
  status: RecipientStatus;
  /** @nullable */
  error?: string | null;
}