# lib/api-zod/src/generated/types/recipientBulkInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MailFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecipientRow } from './recipientRow';

export interface RecipientBulkInput {
  recipients: RecipientRow[];
}